+++ src/Responses/ProductPrice.php
@@ -25,16 +25,16 @@ class ProductPrice
     public function __construct($data)
     {
         if (isset($data->years)) {
-            $this->years = (int)$data->years;
+            $this->years = (int) $data->years;
         }
         if (isset($data->price)) {
-            $this->price = (float)$data->price;
+            $this->price = (float) $data->price;
         }
         if (isset($data->extraPrice)) {
-            $this->priceExtra = (float)$data->extraPrice;
+            $this->priceExtra = (float) $data->extraPrice;
         }
         if (isset($data->extraPriceWildcard)) {
-            $this->priceExtraWildcard = (float)$data->extraPriceWildcard;
+            $this->priceExtraWildcard = (float) $data->extraPriceWildcard;
         }
     }
 }